Organize all ListCommand into a package
Improves code readability and folder structure.
