+++ src/main/java/seedu/address/logic/commands/listcommands/ListCompaniesCommand.java
@@ -1,15 +1,16 @@
-package seedu.address.logic.commands;
+package seedu.address.logic.commands.listcommands;
 
 import static seedu.address.model.Model.PREDICATE_SHOW_ALL_PERSONS;
 
+import seedu.address.logic.commands.CommandResult;
 import seedu.address.model.Model;
 
 /**
- * Lists all persons in the address book to the user.
+ * Lists all companies in the address book to the user.
  */
-public class ListPersonsCommand extends ListCommand {
+public class ListCompaniesCommand extends ListCommand {
 
-    public static final String MESSAGE_SUCCESS = "Listed all persons";
+    public static final String MESSAGE_SUCCESS = "Listed all companies";
 
     @Override
     public CommandResult execute(Model model) {
