+++ src/test/java/seedu/address/logic/LogicManagerTest.java
@@ -20,9 +20,9 @@ import org.junit.jupiter.api.io.TempDir;
 
 import seedu.address.logic.commands.AddCommand;
 import seedu.address.logic.commands.CommandResult;
-import seedu.address.logic.commands.ListCommand;
-import seedu.address.logic.commands.ListPersonsCommand;
 import seedu.address.logic.commands.exceptions.CommandException;
+import seedu.address.logic.commands.listcommands.ListCommand;
+import seedu.address.logic.commands.listcommands.ListPersonsCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
